Move BackendUtils to opt
