+++ compiler/src/dotty/tools/backend/jvm/opt/ClosureOptimizer.scala
@@ -24,7 +24,7 @@ import scala.tools.asm.tree.*
 import dotty.tools.dotc.core.Decorators.em
 import dotty.tools.dotc.util.NoSourcePosition
 import dotty.tools.backend.jvm.BTypes.InternalName
-import dotty.tools.backend.jvm.BackendUtils.*
+import BackendUtils.*
 import dotty.tools.backend.jvm.analysis.{AsmAnalyzer, ProdConsAnalyzer}
 import BCodeUtils.*
